Moved cross staking implementation to main contract file
