+++ contracts/external-staking/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod contract;
-mod cross_staking_api;
 pub mod error;
 pub mod msg;
 pub mod state;
